fix(blog): renamed ramblings to thoughts, revised nested page structure
