--- pages/b_index
@@ -1,15 +1,0 @@
-<template>
-  <BlogList :posts="posts" />
-</template>
-
-<script>
-import BlogList from '~/components/blog/BlogList.vue'
-
-export default {
-  name: 'AllPostsList',
-  components: { BlogList },
-  props: {
-    posts: { type: Array, required: false, default: () => [] },
-  },
-}
-</script>
