+++ pages/thoughts/_slug.vue
